Ai Chatbot/AiChatbot/pyproject.toml
[project]
name = "repl-nix-simpleaichatbot"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.40.0",
    "fastapi>=0.115.6",
    "google-generativeai>=0.8.3",
    "huggingface-hub>=0.26.5",
    "langgraph>=0.2.56",
    "openai>=1.57.0",
    "python-multipart",
    "requests>=2.32.3",
    "sse-starlette>=2.1.3",
    "streamlit>=0.0.19",
    "uvicorn>=0.32.1",
]
